src-tauri/requirements.txt
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4 